EE-9537 - Split out the name combination generator
